merge

+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/conf/graph/GraphVertex.java
@@ -31,8 +31,8 @@ import org.nd4j.shade.jackson.annotation.JsonTypeInfo;
 import java.io.Serializable;
 
 /**
- * A GraphVertex is a vertex in the computation graph. It may contain Layer, or define some arbitrary forward/backward pass
- * behaviour based on the inputs
+ * A GraphVertex is a vertex in the computation graph type of neural network. It may contain Layer, or define some
+ * arbitrary forward/backward pass behaviour based on the inputs. GraphVertex instances may also have trainable parameters.
  *
  * @author Alex Black
  */
@@ -49,7 +49,7 @@ public abstract class GraphVertex implements Cloneable, Serializable {
     @Override
     public abstract int hashCode();
 
-    public abstract int numParams(boolean backprop);
+    public abstract long numParams(boolean backprop);
 
     /**
      * @return The Smallest valid number of inputs to this vertex
