+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/conf/graph/SubsetVertex.java
@@ -33,7 +33,7 @@ import java.util.Arrays;
  * For example, a subset of the activations out of a layer.<br>
  * Note that this subset is specifying by means of an interval of the original activations.
  * For example, to get the first 10 activations of a layer (or, first 10 features out of a CNN layer) use
- * new SubsetVertex(0,9).<br>
+ * {@code new SubsetVertex(0,9)}<br>
  * In the case of convolutional (4d) activations, this is done along channels.
  *
  * @author Alex Black
@@ -72,7 +72,7 @@ public class SubsetVertex extends GraphVertex {
     }
 
     @Override
-    public int numParams(boolean backprop) {
+    public long numParams(boolean backprop) {
         return 0;
     }
